Change floats to doubles and work around a firefox bug.
1) Floats and doubles are equivalent in JS but using floats requires casting (e.g., 3.14 -> 3.14f or Math.PI -> (float)Math.PI). Therefore, I've switched to just using doubles.

2) Firefox has a bug where the anticlockwise parameter to arc is not optional (see https://bugzilla.mozilla.org/show_bug.cgi?id=617319). I've worked around it and have added a test case.

Review at http://gwt-code-reviews.appspot.com/1192801

Review by: rice@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@9369 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ user/src/com/google/gwt/canvas/dom/client/CanvasGradient.java
@@ -40,7 +40,7 @@ public class CanvasGradient extends FillStrokeStyle {
    * @param offset value between 0 and 1 for where the color stop is located.
    * @param color color at the stop.
    */
-  public final native void addColorStop(float offset, String color) /*-{
+  public final native void addColorStop(double offset, String color) /*-{
     this.addColorStop(offset, color);
   }-*/;
 }
